requirements.txt
django==4.1.3
python-dotenv==0.21.0
psycopg2-binary==2.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-filter==22.1
celery==5.2.7
Redis==4.3.4